Moving AsyncCommandQueue to serial/.
