+++ iree/hal/host/serial/CMakeLists.txt
@@ -16,6 +16,43 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
+    async_command_queue
+  HDRS
+    "async_command_queue.h"
+  SRCS
+    "async_command_queue.cc"
+  DEPS
+    absl::core_headers
+    absl::synchronization
+    iree::base::status
+    iree::base::tracing
+    iree::hal::command_queue
+    iree::hal::host::serial::serial_submission_queue
+    iree::hal::semaphore
+  PUBLIC
+)
+
+iree_cc_test(
+  NAME
+    async_command_queue_test
+  SRCS
+    "async_command_queue_test.cc"
+  DEPS
+    ::async_command_queue
+    absl::memory
+    absl::time
+    iree::base::status
+    iree::base::status_matchers
+    iree::base::time
+    iree::hal::command_queue
+    iree::hal::host::serial::serial_submission_queue
+    iree::hal::testing::mock_command_buffer
+    iree::hal::testing::mock_command_queue
+    iree::testing::gtest_main
+)
+
+iree_cc_library(
+  NAME
     serial_command_processor
   HDRS
     "serial_command_processor.h"
@@ -41,6 +78,7 @@ iree_cc_library(
   SRCS
     "serial_scheduling_model.cc"
   DEPS
+    ::async_command_queue
     ::serial_command_processor
     ::serial_submission_queue
     absl::inlined_vector
@@ -48,7 +86,6 @@ iree_cc_library(
     iree::base::source_location
     iree::base::status
     iree::base::tracing
-    iree::hal::host::async_command_queue
     iree::hal::host::host_semaphore
     iree::hal::host::inproc_command_buffer
     iree::hal::host::nop_event
