+++ iree/hal/host/serial/serial_scheduling_model.cc
@@ -15,10 +15,10 @@
 #include "iree/hal/host/serial/serial_scheduling_model.h"
 
 #include "iree/base/tracing.h"
-#include "iree/hal/host/async_command_queue.h"
 #include "iree/hal/host/host_semaphore.h"
 #include "iree/hal/host/inproc_command_buffer.h"
 #include "iree/hal/host/nop_event.h"
+#include "iree/hal/host/serial/async_command_queue.h"
 #include "iree/hal/host/serial/serial_command_processor.h"
 #include "iree/hal/host/serial/serial_submission_queue.h"
